Arreglado problema de update planes

+++ app/Console/Commands/UpdatePlanes.php
@@ -46,15 +46,17 @@ class UpdatePlanes extends Command
 
 
         foreach ($planUser as $pivot) {
-            if (!$pivot->user->isInactivo()){
-                if($pivot->vencePorFecha() && $pivot->isVencido()) {
-                    $pivot->adeudar();
-                    $pivot->renovar();
-                    if ($pivot->deuda) {
-                        if ($pivot->user->hasFamilia()){
-                            $pivot->user->familia->deudas()->save($pivot->deuda);
-                        } else {
-                            $pivot->user->deudas()->save($pivot->deuda);
+            if ($pivot->user->exists()) {
+                if (!$pivot->user->isInactivo()) {
+                    if ($pivot->vencePorFecha() && $pivot->isVencido()) {
+                        $pivot->adeudar();
+                        $pivot->renovar();
+                        if ($pivot->deuda) {
+                            if ($pivot->user->hasFamilia()) {
+                                $pivot->user->familia->deudas()->save($pivot->deuda);
+                            } else {
+                                $pivot->user->deudas()->save($pivot->deuda);
+                            }
                         }
                     }
                 }
@@ -64,18 +66,20 @@ class UpdatePlanes extends Command
         $especialUser = EspecialUser::where('pagado','=',1)->with('user')->get();
 
         foreach ($especialUser as $pivot) {
-            if (!$pivot->user->isInactivo()) {
-                if ($pivot->vencePorFecha() && $pivot->isVencido() && ($pivot->especial->renovable == 1)) {
-                    $pivot->adeudar();
-                    if ($pivot->user->hasFamilia()) {
-                        $pivot->user->familia->deudas()->save($pivot->deuda);
-                    } else {
-                        $pivot->user->deudas()->save($pivot->deuda);
+            if ($pivot->user->exists()) {
+                if (!$pivot->user->isInactivo()) {
+                    if ($pivot->vencePorFecha() && $pivot->isVencido() && ($pivot->especial->renovable == 1)) {
+                        $pivot->adeudar();
+                        if ($pivot->user->hasFamilia()) {
+                            $pivot->user->familia->deudas()->save($pivot->deuda);
+                        } else {
+                            $pivot->user->deudas()->save($pivot->deuda);
+                        }
+                        $pivot->renovar();
+                    } elseif ($pivot->isVencido() && ($pivot->especial->renovable == 0)) {
+                        $pivot->especial()->delete();
+                        $pivot->delete();
                     }
-                    $pivot->renovar();
-                } elseif ($pivot->isVencido() && ($pivot->especial->renovable == 0)) {
-                    $pivot->especial()->delete();
-                    $pivot->delete();
                 }
             }
         }
